requirements.txt
# Bibliothèques pour l'environnement et les variables d'environnement
python-dotenv

# Génération de texte et images
huggingface_hub
google-generativeai
Pillow

# Similarité de texte
fuzzywuzzy

# Serveur web
flask

# Requêtes HTTP
requests

# Traitement XML
#xml.etree.ElementTree
# Pour assurer la génération de références, prompts, et images
# Ajoutez les dépendances des fichiers de modules si elles ne sont pas incluses ici.

# Ajout de dépendances supplémentaires pour votre application
